requirements.txt
opencv-python
cv2-enumerate-cameras
pyside6
numpy
vosk
pyboof
piper-tts
pygame
sounddevice
python-Levenshtein
PyYAML
